UPSTREAM: HID: sony: Remove the size check for the Dualshock 4 HID Descriptor
Sony has modified the HID descriptor in new revisions of the Dualshock 4 which
causes the size check in the descriptor replacement function to fail.  Remove it
so that new revisions of the controller will work correctly.

The module is completely replacing the descriptor instead of patching it, so the
size check isn't really necessary anyways.

Signed-off-by: Frank Praznik <frank.praznik@gmail.com>
Signed-off-by: Jiri Kosina <jkosina@suse.cz>
(cherry picked from commit b71b5578a84d297954e4812ba0ca2d466e61cf42)

Bug: 111431828
Signed-off-by: Kim Low <kim-huei.low@sony.com>
Change-Id: I49f5c7ea2b50394fa6e75c4da161f6d5dfd82072
Signed-off-by: Siarhei Vishniakou <svv@google.com>