fs/pstore/ramoops: add alternate encrypted buffer support
This patch allows the ramoops to read previous logs from an
AES encrypted location while still allowing writes to the
current log.

This is used to allow the backing store to be encrypted and
then written to disk by the bootloader.  The keys to decrypt
are loaded by userspace and then pstore is remounted to
force a reload of the now decrypted backing store.

Bug: 37554629
Bug: 37553996
Change-Id: Ieb7615d22bdfbf2130a0f6ca1b4038ffdf2fe338
Signed-off-by: Patrick Tjin <pattjin@google.com>