qpnp-smbcharger: fix slower charging with type-C charger
Charger is misjudged as bad DCP charger(5V1A) and force
limit current to 1A if two dischg time is very close when
doing charging stress test(connect for 10 seconds to a
charger and then disconnect for 3 seconds).

The check WA is for 5V/1A DCP only. Don't apply it for type-C
charger to fix it.

Bug: 113164315
Change-Id: Ic49c026d401c346721ea7808cde6410bf2d5824c
Signed-off-by: Jack Wu <wjack@google.com>