kernel: Allow arbitrary timer frequency values
Rather than manually adding an option for each individual timer
frequency that may be desired, make the config option variable instead.
This allows for arbitrary frequencies within the range that the kernel
is able to be compiled with, taken from jiffies.h. Many of the possible
values are not sane and may not boot, so take care in the values chosen.
The config option is left open-ended for experimenting, but that does
not mean that it is wise to pick values such as 5000.

Note that this may break drivers which rely on directly checking the
CONFIG_HZ_NNN options (e.g. sched/walt.c since commit 578db5dbecfa), so
be sure to fix such occurrences.

Test: kernel compiles, device boots, exposed frequency is updated
Signed-off-by: Tyler Nijmeh <tylernij@gmail.com>
Signed-off-by: kdrag0n <dragon@khronodragon.com>