Add performance profiles and use magisk for it
"SEPolicy blocks init.rc from writing anything directly from it.
But, the init.rc can setprop and execute from seperate scripts.
When the intent from spectrum is processed via the init.rc, the
init.rc pushes the command to execute the init.sh for the respective
profile. With time, I can combine the init.sh files, but for now, this
works"

This is a proof-of-concept. anything in these files are not gonna be used
in my kernel

Signed-off-by: kingbri <bdashore3@gmail.com>