Squash kdrag0n commits
kernel: Only boost on fork if the screen is on

This prevents unnecessary boosting when services and/or content
providers start in the background to perform jobs while the device is
not actively being used.

kernel: Boost DDR bus on zygote fork

No clue how this was left out in the end.

Signed-off-by: kdragon <dragon@khronodragon.com>
Signed-off-by: kingbri <bdashore3@gmail.com>