defconfig: Enable hang detection
Same as watchdog

The sultan logic: "It may be rare that we get a lockup, but what if it happens in the
middle of the night and your alarm doesn't fire because of it? According
to Sultan, even the tiniest of embedded devices have watchdogs. It fires
rarely enough that it really doesn't matter."

Signed-off-by: kdrag0n <dragon@khronodragon.com>
Signed-off-by: kingbri <bdashore3@gmail.com>