cpu_input_boost: add display state negative stune boosting
"This commit allows for schedtune boosting to dip below zero when
the screen is off. It applies a set value of -20 to all schedtune boosts
Doing this provides further battery savings
especially when using background apps such as YT music. It is required
to have Always On Display off to experience the full power
when using this commit because fb_notifier will not register as
screen_off if AOD is used."

consolidated all boosts from kdragon's commit

https://github.com/kdrag0n/proton_bluecross/commit/83d73902a191da47aee11aa61f4bbaa30522eb22

Signed-off-by: kingbri <bdashore3@gmail.com>