power: Introduce CorePower system power state optimizer
CorePower is a driver to adapt the system in response to events that
signal a power state transition. This can help reduce power consumption
in certain states by performing actions such as CPU isolation that trade
performance for reduced power usage in states where such trade-offs are
appropriate.

Currently, the driver is able to perform the following actions:
 - Force cpuidle to select the deepest CPU idle state available whenever
 it decides to enter idle
 - Force lpm-levels to select the deepest CPU cluster idle state
 available whenever it decides to enter idle

All aspects of the driver are dynamically configurable via module
parameters, which can be modified at boot time through kernel command
line arguments or at runtime in sysfs.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>