cpufreq: Only block setting the max freq for init and libperfmgr
Now that we have improved msm-thermal-simple to deal with non-standard
max frequencies properly, unlock this setting to userspace processes.
init and libperfmgr remain blocked because their settings may override
user choices and are unnecessary with msm-thermal-simple.

This makes it possible for users to use their max frequency of choice
instead of forcing 2323200 KHz on everyone.

Also fix the indentation while we're at it.

Signed-off-by: Danny Lin <danny@kdrag0n.dev>