Don't use WALT_CPU_UTIL
"I have observed that walt_cpu_util has been causing the freqs to
jump to max then min with no in-between. An xda member by the name
levan_i suggested that I disable this and the freq jumps stopped
immediately. Using PELT for cpu will be a good idea from now on"

Signed-off-by: kingbri <bdashore3@gmail.com>