Corepower: Add interface for negative display state aware suspend boosting
"This creates a suspend boost interface which sets the root schedtune boost
to -50 when screen is off and clears it when screen turns on. Using
this has reduced my idle drain by about half of what it used to be when
listening to YT music for approximately 5 songs in succession using
background play. This also works with AOD because fb_notifier does not
have a low power setting and registers the display as off

The feature can be turned off using a module in path:
/sys/module/corepower/parameters/suspend_boost_enable

Also, the boost value can be changed in the path:
/sys/module/corepower/parameters/suspend_stune_boost"

Signed-off-by: Brian Dashore <bdashore3@gmail.com>