cpu_input_boost: Block max boosts while the screen is off
In the majority of cases, we actually *do* want to block max boost kicks
if the screen is off. The kicks that we want to remain functional happen
to all be wake boosts, so allow external wake boosts to continue to
function. All other max boosts are now ineffectual.

Modified to use new sultan commits in cpu_input_boost and wake_boost
doesn't exist in my kernel

Signed-off-by: Danny Lin <danny@kdrag0n.dev>
Signed-off-by: kingbri <bdashore3@gmail.com>